% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ABtest}
\alias{ABtest}
\title{A-B testing dataset}
\format{
A data frame with 100 rows and 6 variables; each row represents
a visitor to a company's mobile website.
\describe{
  \item{UserID}{A unique identifier for the dataset}
  \item{OperatingSystem}{User's mobile operating system, Android or iOS}
  \item{UserLocation}{Country in which the user is located, Canada or USA}
  \item{WebsiteVersion}{Version of the mobile website displayed to the user, A or B}
  \item{Clicked}{Binary indicator for whether the user clicked on a link on the website, Yes or No}
  \item{SecondsOnSite}{Total time in seconds that the user spent on the company's home page}
  }
}
\usage{
ABtest
}
\description{
A dataset containing simulated data for a fictional company that is performing
an A-B test to evaluate two versions of its mobile website.
}
\examples{
data(ABtest)
}
\keyword{datasets}
