% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BostonWeather}
\alias{BostonWeather}
\title{Boston weather dataset}
\format{
A data frame with 95 rows and 8 variables; each row represents a month.
\describe{
  \item{t}{A time index numbering the periods in the data set, 1-95}
  \item{year}{Calendar year, 2014-2021}
  \item{month}{Calendar month, January-December}
  \item{avg.temp}{Average temperature reading in degrees Fahrenheit (ºF)}
  \item{precip}{Total of all types of precipitation recorded in the month in inches}
  \item{snow}{Total snow recorded in the month in inches}
  \item{max.temp}{Maximum temperature recorded in the month in degrees Fahrenheit (ºF)}
  \item{min.temp}{Minimum temperature recorded in the month in degrees Fahrenheit (ºF)}
  }
}
\usage{
BostonWeather
}
\description{
A dataset containing monthly temperature and precipitation data
from a weather station in the city of Boston, MA.
}
\details{
#' @source \url{https://www.weather.gov/wrh/Climate}
}
\examples{
data(BostonWeather)
}
\keyword{datasets}
