% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{customers}
\alias{customers}
\title{Retail case study data set: Customers}
\format{
A data frame with 2501 rows and 7 variables:
\describe{
  \item{customer.id}{A unique identifier for the customers in the store's loyalty program (an alphanumeric code)}
  \item{customer.state}{State in which the customer's registered address is located}
  \item{age}{Customer's age, based on reported birthday at sign up}
  \item{birthday.month}{Numeric code for month of customer's birthday, used to send a promotional coupon (1 = January, 2 = February, etc.)}
  \item{years.as.member}{Number of full years the customer has been part of the loyalty program, 0 if less than 1 year}
  \item{in.store.exp}{On a customer survey distributed to all loyalty program members, customer's response to the question: "On a scale of 1-5, how satisfied are you with the IN-STORE SHOPPING EXPERIENCE at our stores?" Non-responders will have a blank (missing) value.}
  \item{selection}{On a customer survey distributed to all loyalty program members, customer's response to the question: "On a scale of 1-5, how satisfied are you with the SELECTION OF PRODUCTS at our stores?" Non-responders will have a blank (missing) value.}
  }
}
\source{
Created for academic use.
}
\usage{
customers
}
\description{
Synthetic data created to reflect the trends and relationships
in an actual retail data set. This table can be linked to
'stores' and 'sales.' Each row represents one member of the store's
customer loyalty program.
}
\examples{
data(customers)
}
\keyword{datasets}
