% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rides}
\alias{rides}
\title{Bluebikes rides dataset}
\format{
A data frame with 1275 rows and 10 variables; each row represents a ride.
\describe{
  \item{ride.id}{A unique identifier for the dataset}
  \item{ride.date}{Calendar date of the ride, defined based on ride start time}
  \item{ride.weekday}{Day of week of the ride}
  \item{start.time}{Start date and time of the ride}
  \item{end.time}{End date and time of the ride}
  \item{rider.type}{Type of rider taking the ride, 'casual' or 'member'
  where 'casual' indicates a one-time purchase; 'member' indicates the rider has
  a monthly or annual membership}
  \item{duration.min}{Duration of the ride in minutes}
  \item{start.station.id}{Identification code for the station where the ride was started.
  All rides must begin and end at a station.}
  \item{end.station.id}{Identification code for the station where the ride was started}
  \item{straight.line.dist}{Straight line distance between start station and end station,
  calculated based on their latitude and longitude. Note that rides may begin and end at
  the same station.}
  }
}
\source{
\url{https://www.bluebikes.com/system-data}
}
\usage{
rides
}
\description{
A dataset containing time, location, and other attributes for
a subset of rides taken in Bluebikes bike share network in the
Boston, MA metro area.
}
\examples{
data(rides)
}
\keyword{datasets}
