% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sales}
\alias{sales}
\title{Retail case study data set: Sales}
\format{
A data frame with 5961 rows and 15 variables:
\describe{
  \item{store}{The retail store where the sale took place (an alphanumeric code)}
  \item{sale.date}{The date of the sale}
  \item{tran.type}{Type of transaction; a categorical variable that may be equal to "Sale," "Return," or "Adjustment"}
  \item{tran.number}{A numeric code used by the retailer to uniquely identify distinct transactions across the chain (transactions may contain multiple sales records)}
  \item{tran.item.number}{Counter indicating a sale record's item number within its transaction, e.g. 1 for the first SKU scanned, 2 for the second SKU scanned, etc.}
  \item{sku}{A numeric code used to identify each unique product sold in the stores. A given product (e.g., Sierra Jacket) will have multiple SKUs corresponding to each size and color offered.}
  \item{category}{Product category of the item represented in the sales record. Legitimate values include:Backpacks, Bottoms, Camping Misc, Footwear, Jackets, Sleeping Gear, Swimwear, Tops, Travel.}
  \item{qty}{Number of identical items purchased in the same sales record}
  \item{unit.cost}{Cost per unit of the inventory represented in the sale record. Note that the relevant total cost for a sale record row is qty*unit.cost.}
  \item{sale.amount}{Total revenue collected for the sale record, reflecting the actual price paid by the customer with discounts and coupons.}
  \item{unit.original.retail}{Original retail price per unit for the inventory represented in the sale record.}
  \item{gross.margin}{Gross margin of the sale record, calculated as (sale.amount - qty*unit.cost)/sale.amount}
  \item{price.category}{Price category of the inventory reflected in the sale record row. Legitimate values are Full Price, Markdown, Coupon, and Clearance.}
  \item{loyalty.member}{A binary indicator of whether the customer making the purchase is a member of the store's loyalty program. Acceptable values are 0 and 1.}
  \item{customer.id}{A unique identifier given to each customer who registers with the store loyalty program. If customer did not enter a loyalty card number, C99999 is shown.}
  }
}
\source{
Created for academic use.
}
\usage{
sales
}
\description{
Synthetic data created to reflect the trends and relationships
in an actual retail data set. This table can be linked to
'stores' and 'customers.' Each row represents one sales record from
the store's retail outlets.
}
\examples{
data(sales)
}
\keyword{datasets}
