% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stores}
\alias{stores}
\title{Retail case study data set: Stores}
\format{
A data frame with 5 rows and 5 variables:
\describe{
  \item{store.id}{A unique identifier for the stores in the retail chain (a numeric code)}
  \item{city}{City in which the store is located}
  \item{state}{U.S. state in which the store is located}
  \item{sqft}{Total square footage of the store location}
  \item{store.cat}{Numeric code indicating the sales volume tier to which the store belongs (1 = Largest, 3 = Smallest)}
  }
}
\source{
Created for academic use.
}
\usage{
stores
}
\description{
Synthetic data created to reflect the trends and relationships
in an actual retail data set. This table can be linked to
'customers' and 'sales.' Each record represents one of the retail
locations included in the study.
}
\examples{
data(stores)
}
\keyword{datasets}
